Rename the "winx" crate to "windows".
"winx" is now just one of multiple windows API crates we use, and
"windows" is the more common in Rust code bases.
