+++ cap-primitives/src/windows/fs/dir_entry_inner.rs
