+++ cap-primitives/src/windows/fs/dir_options_ext.rs
