+++ cap-primitives/src/windows/fs/errors.rs
