+++ cap-primitives/src/windows/fs/get_path.rs
