+++ cap-primitives/src/windows/fs/is_read_write_impl.rs
