+++ cap-primitives/src/windows/fs/oflags.rs
