+++ cap-primitives/src/windows/fs/open_options_ext.rs
