+++ cap-primitives/src/windows/fs/open_unchecked.rs
