+++ cap-primitives/src/windows/fs/read_dir_inner.rs
