+++ cap-primitives/src/windows/fs/read_link_impl.rs
