+++ cap-primitives/src/windows/fs/remove_dir_all_impl.rs
