+++ cap-primitives/src/windows/fs/remove_dir_unchecked.rs
