+++ cap-primitives/src/windows/fs/remove_file_unchecked.rs
