+++ cap-primitives/src/windows/fs/remove_open_dir_impl.rs
