+++ cap-primitives/src/windows/fs/stat_unchecked.rs
