+++ cap-primitives/src/windows/fs/symlink_unchecked.rs
